Inventario Ordenes de trabajo listo, falta terminar de darle diseño

+++ LubriTech/Controller/PriceList_Controller.cs
@@ -46,7 +46,7 @@ namespace LubriTech.Controller
             return new Prices_Model().deletePrice(id);
         }
 
-        public decimal getPriceByItem(string itemCode, int priceListId)
+        public double getPriceByItem(string itemCode, int priceListId)
         {
             return new Prices_Model().getPriceByItem(itemCode, priceListId);
         }
